reorganized railties, added coder default, added coder test
